ntfy/requirements.txt
flask
apprise
python-dotenv
